+++ include/optionx_cpp/data/trading/TradeRecordFilter.hpp
@@ -77,7 +77,7 @@ namespace optionx {
         IncludeExcludeFilter<std::int64_t> account_ids;
         IncludeExcludeFilter<std::int64_t> option_ids;
         IncludeExcludeFilter<std::uint32_t> trade_ids;
-        IncludeExcludeFilter<std::uint64_t> signal_ids;
+        IncludeExcludeFilter<std::uint32_t> signal_ids;
         IncludeExcludeFilter<std::int64_t> request_unique_ids;
 
         IncludeExcludeFilter<std::int64_t> durations;
